# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  domains.Error:
    properties:
      error:
        example: Error message
        type: string
    type: object
...
  domains.SkinType:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API for managing products, categories, brands, and skin types in an
    e-commerce system
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: E-commerce API
  version: "1.0"
paths:
...
  /skin-types:
    get:
      consumes:
      - application/json
      description: Get a list of all skin types
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domains.SkinType'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domains.Error'
      summary: Get all skin types
      tags:
      - skin-types
    post:
      consumes:
      - application/json
      description: Create a new skin type with the provided details
      parameters:
      - description: Skin Type object
        in: body
        name: skinType
        required: true
        schema:
          $ref: '#/definitions/domains.SkinType'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domains.SkinType'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domains.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domains.Error'
      summary: Create a new skin type
      tags:
      - skin-types
  /skin-types/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a skin type by its ID
      parameters:
      - description: Skin Type ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domains.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domains.Error'
      summary: Delete skin type
      tags:
      - skin-types
    get:
      consumes:
      - application/json
      description: Get a skin type by its ID
      parameters:
      - description: Skin Type ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domains.SkinType'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domains.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domains.Error'
      summary: Get skin type by ID
      tags:
      - skin-types
    put:
      consumes:
      - application/json
      description: Update an existing skin type
      parameters:
      - description: Skin Type ID
        in: path
        name: id
        required: true
        type: integer
      - description: Skin Type object
        in: body
        name: skinType
        required: true
        schema:
          $ref: '#/definitions/domains.SkinType'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domains.SkinType'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domains.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domains.Error'
      summary: Update skin type
      tags:
      - skin-types
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"